+++ CHANGELOG.rst
@@ -9,6 +9,10 @@
 * revise README [jcupitt]
 * add watermark example [jcupitt]
 * fix syntax highlighting in README [favorable-mutation]
+* add signal handling [jcupitt]
+* add perf tests [kleisauke]
+* speed up Operation.call [kleisauke]
+* fix logging [h4l]
 
 ## Version 2.1.8 (1 Jul 2019)